+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 --------------------------------------------------------------------------------------------------
 
 * Adjust rendering of BitBar plugin errors.
+* Fix fetching of log level names in Python 3.4+.
 
 
 `2.0.0 (2016-12-04) <https://github.com/kdeldycke/meta-package-manager/compare/v1.12.0...v2.0.0>`_